+++ bonus/srcs/graphics/raycasting/draw/sprites/set_sprites_attr.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   set_sprites_attr.c                                 :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: psalame <psalame@student.42angouleme.fr    +#+  +:+       +#+        */
+/*   By: edbernar <edbernar@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/18 14:16:42 by psalame           #+#    #+#             */
-/*   Updated: 2024/04/18 17:30:57 by psalame          ###   ########.fr       */
+/*   Updated: 2024/04/19 13:02:07 by edbernar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,7 +33,7 @@ void	set_sprites_screen_x(t_sprite *sprites, t_vec3 plyPos, t_mlx *mlx)
 		transform = inv_cam * (dir->y * diff.x - dir->x * diff.y);
 		sprites->depth = inv_cam * (plane->x * diff.y - plane->y * diff.x);
 		if (sprites->depth == 0.0f)
-			sprites->depth = INT_MAX;
+			sprites->depth = (float)INT_MAX;
 		sprites->screen_x = (mlx->stg->width / 2)
 			* ((1 + transform / sprites->depth));
 		sprites++;
